refactor: split work-item claim discovery from allocator (INFRA-179)
Work-Run: d3f0023b-9336-4c18-ba11-1601673b77c6
Work-Receipt: g0-r0
